requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
djoser==2.2.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
setuptools==68.2.2
django-filter==24.2
drf-spectacular==0.24.1